test now fails if user isn't logged in can create
The test was passing even though my code was allowing a non-logged in user to create a note and save it to the database.  The user_id column was nil, but the note was created nonetheless.  The added assertion makes sure that the content of the last note of the database doesn't match the content of the note that was submitted to the post request to the NotesController's :create action.